NDI-86: Consolidate all page objects into one
