+++ backstop_data/engine_scripts/puppet/common/open-accordions.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Page = require('../page-objects/accordion-page.js');
+const Page = require('../page-objects/crm-page.js');
 
 module.exports = async (engine, scenario, vp) => {
   const page = new Page(engine, scenario, vp);
