--- backstop_data/engine_scripts/puppet/page-objects/accordion-page.js
@@ -1,13 +1,0 @@
-'use strict';
-
-var FormPage = require('./form-page.js');
-
-module.exports = class AccordionPage extends FormPage {
-
-  /**
-   * Opens all the accordions on the page
-   */
-  async openAllAccordions() {
-    await this.clickAll('div.crm-accordion-wrapper.collapsed > div');
-  }
-};
